feat(ti): Introduce PSCI Driver for AM62L
The AM62L is a lite, low power and performance optimized family of
application processors that are built for Linux application development.

There is enough deviation from the previous K3 gen SoCs with regards to how the
PSCI functionality looks like on this device. For example, it no longer does
reset or LPSC turn ON/OFF operations using any external Device Management
entity like it did earlier.
The actual power ON/OFF operations will be implemented in a later stage, once
all the PM related drivers are integrated. Such places in this driver have been
marked as TODO and will be replaced with actual call to power on/off the cores
at a later point.

This warranted the addition of a new am62l_psci driver for the AM62L family of
devices. For further details of how this new device looks like, refer to the
TRM [1].

[1] https://www.ti.com/lit/pdf/sprujb4

Change-Id: Ic53096e7bbc25fa55386ac4b6bef364dd6d0cf3b
Signed-off-by: Dhruva Gole <d-gole@ti.com>